HOCS-1076 Use cert-manager.io
The Kubernetes configuration prior to this commit used kube-cert-manager
to manage certificates from Let's Encrypt. However, kube-cert-manager is
getting deprecated upstream and will cease working in June 2020.

This commit updates the configuration to use Jetstack's cert-manager,
which is now deployed on the ACP clusters.

The TLS keypair is stored in a secret. In similar commits to this one elsewhere
in the HOCS ecosystem, we rename the secret to prevent multiple certificate
managers managing the same certificate. We don't need to bother here, as
there's no production to worry about.