Lower proxy read timeout
For unidentified reasons some of the CSS in the test reports is timing
out instead of rendering. Because external CSS stylesheets are
render-blocking resources, this problem causes reports to not be
rendered for the full 2 minutes before the .css file times out.

Instead of actually fixing the problem, this commit lowers the timeout
from the default of 2 minutes to just 10 seconds. This means that
reports still take a while to load, but means you don't have to wait too
long.